# make `T.let` faster for simple module types