### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

`T.let` with a simple `Module` type (`Integer`, `String`, etc.) is moderately common.  We can speed up this case by not fetching an entire `T::Types::Simple` structure from the cached pool, but instead check `is_a?` directly on such types for casts.

Benchmarks before (non-`let` benchmarks elided):

```
froydnj@pyro:~/src/sorbet/gems/sorbet-runtime$ env PATH=$PATH:/home/froydnj/ruby/bin /home/froydnj/ruby/bin/bundle exec rake bench:typecheck
T.let(..., Integer): 334.068 ns
T.let(..., T.nilable(Integer)): 834.552 ns
T.let(..., Example): 344.172 ns
T.let(..., T.nilable(Example)): 839.652 ns
```

Benchmarks after (nilable benchmarks not expected to change much and are within noise for these benchmarks):

```
T.let(..., Integer): 193.649 ns
T.let(..., T.nilable(Integer)): 825.172 ns
T.let(..., Example): 187.888 ns
T.let(..., T.nilable(Example)): 824.67 ns
```

Good for a 40%-ish speedup, which is nice.

### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

See included automated tests.